confirmação de exclusão

+++ uesb_forms/lib/Componentes/BancoDeQuestoes/widgetBanco_Questao.dart
@@ -4,12 +4,11 @@ import 'package:uesb_forms/Controle_Modelo/banco_list.dart';
 import 'package:uesb_forms/Modelo/Banco.dart';
 import 'package:uesb_forms/Utils/rotas.dart';
 
-
 class WidgetbancoQuestao extends StatelessWidget {
   final Banco banco;
-  
+
   const WidgetbancoQuestao({
-    super.key, 
+    super.key,
     required this.banco,
   });
 
@@ -60,15 +59,49 @@ class WidgetbancoQuestao extends StatelessWidget {
                           icon: Icon(Icons.delete),
                           color: Color.fromARGB(255, 27, 7, 80),
                           onPressed: () async {
-                            try {
-                              await bancoList.excluirBanco(banco.id ?? '');
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                SnackBar(content: Text('Banco excluído com sucesso')),
-                              );
-                            } catch (e) {
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                SnackBar(content: Text('Erro ao excluir o banco')),
-                              );
+                            // Exibe um diálogo de confirmação
+                            final confirm = await showDialog<bool>(
+                              context: context,
+                              builder: (BuildContext context) {
+                                return AlertDialog(
+                                  title: Text('Confirmar Exclusão'),
+                                  content: Text(
+                                      'Tem certeza de que deseja excluir este banco?'),
+                                  actions: <Widget>[
+                                    TextButton(
+                                      child: Text('Cancelar'),
+                                      onPressed: () {
+                                        Navigator.of(context).pop(
+                                            false); // Retorna 'false' ao pressionar "Cancelar"
+                                      },
+                                    ),
+                                    TextButton(
+                                      child: Text('Excluir'),
+                                      onPressed: () {
+                                        Navigator.of(context).pop(
+                                            true); // Retorna 'true' ao pressionar "Excluir"
+                                      },
+                                    ),
+                                  ],
+                                );
+                              },
+                            );
+
+                            // Se o usuário confirmou a exclusão
+                            if (confirm == true) {
+                              try {
+                                await bancoList.excluirBanco(banco.id ?? '');
+                                ScaffoldMessenger.of(context).showSnackBar(
+                                  SnackBar(
+                                      content:
+                                          Text('Banco excluído com sucesso')),
+                                );
+                              } catch (e) {
+                                ScaffoldMessenger.of(context).showSnackBar(
+                                  SnackBar(
+                                      content: Text('Erro ao excluir o banco')),
+                                );
+                              }
                             }
                           },
                         ),
@@ -79,11 +112,14 @@ class WidgetbancoQuestao extends StatelessWidget {
                             try {
                               await bancoList.duplicarBanco(banco.id ?? '');
                               ScaffoldMessenger.of(context).showSnackBar(
-                                SnackBar(content: Text('Banco duplicado com sucesso')),
+                                SnackBar(
+                                    content:
+                                        Text('Banco duplicado com sucesso')),
                               );
                             } catch (e) {
                               ScaffoldMessenger.of(context).showSnackBar(
-                                SnackBar(content: Text('Erro ao duplicar o banco')),
+                                SnackBar(
+                                    content: Text('Erro ao duplicar o banco')),
                               );
                             }
                           },
